fix(kotlin): resolve all compile errors
- AuthScreens: use Icons.AutoMirrored.Filled.ArrowBack (renamed in Material3)
- FUPScreen: oldFup init with ?: "" to avoid String? -> String mismatch
- FamiliesScreen: onPolicyClick takes Int not Long
- GSTScreen: use planType field (planName doesn't exist on GSTCalculation)
- ViewModels: updateClient/markPaid/updateLead/updateActivity take Int not String
- Screens: add .toInt() when passing String model ids to Int ViewModel params

+++ app/src/main/java/com/balam/crm/ui/screens/ActivitiesScreen.kt
@@ -103,7 +103,7 @@ fun ActivitiesScreen(vm: ActivitiesViewModel = hiltViewModel()) {
             isLoading = createState is UiState.Loading,
             error = (createState as? UiState.Error)?.message,
             onDismiss = { updateActivity = null; vm.resetCreateState() },
-            onConfirm = { status -> vm.updateActivity(activity.id, status) }
+            onConfirm = { status -> vm.updateActivity(activity.id.toInt(), status) }
         )
     }
 }
